requirements.txt
Django==4.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==11.3.0
django-phonenumber-field==8.1.0
phonenumbers==8.13.25
celery==5.3.4
redis==5.0.1
boto3==1.34.0
django-storages==1.14.2
cryptography==44.0.1
mysqlclient==2.2.7
